src/api/datamanage/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile --no-emit-index-url --no-emit-trusted-host requirements.in
#
amqp==2.6.1
    # via
    #   -c ../upizza/requirements.txt
    #   kombu
arrow==1.1.1
    # via jsonext
attrs==20.3.0
    # via
    #   -c ../upizza/requirements.txt
    #   metadata-sdk
    #   pytest
billiard==3.6.4.0
    # via
    #   -c ../upizza/requirements.txt
    #   celery
celery==4.4.7
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
    #   django-celery-beat
    #   django-celery-results
certifi==2020.12.5
    # via
    #   -c ../upizza/requirements.txt
    #   requests
chardet==3.0.4
    # via
    #   -c ../upizza/requirements.txt
    #   requests
concurrent-log-handler==0.9.19
    # via
    #   -c ../upizza/requirements.txt
    #   metadata-sdk
confluent-kafka==1.5.0
    # via -r requirements.in
contextlib2==0.6.0.post1
    # via
    #   -c ../upizza/requirements.txt
    #   opentracing-instrumentation
deepdiff==3.3.0
    # via -r requirements.in
dictdiffer==0.8.1
    # via
    #   -r requirements.in
    #   metadata-sdk
django==2.2.20
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
    #   django-celery-beat
    #   django-filter
    #   django-opentracing
    #   django-redis
    #   django-reversion
    #   django-timezone-field
    #   djangorestframework
django-celery-beat==2.0.0
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
django-celery-results==1.2.1
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django_opentracing==1.1.0
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
django-redis==4.10.0
    # via -r requirements.in
django-reversion==2.0.9
    # via -r requirements.in
django-timezone-field==4.1.2
    # via
    #   -c ../upizza/requirements.txt
    #   django-celery-beat
djangorestframework==3.12.4
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
elasticsearch==6.3.1
    # via -r requirements.in
enum34==1.1.6
    # via -r requirements.in
fastavro==0.14.2
    # via -r requirements.in
future==0.18.2
    # via
    #   -c ../upizza/requirements.txt
    #   opentracing-instrumentation
gevent==20.9.0
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
    #   metadata-sdk
greenlet==1.0.0
    # via
    #   -c ../upizza/requirements.txt
    #   gevent
    #   sqlalchemy
gunicorn==19.6.0
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
httplib2==0.7.3
    # via -r requirements.in
idna==2.8
    # via
    #   -c ../upizza/requirements.txt
    #   requests
importlib-metadata==4.0.0
    # via
    #   -c ../upizza/requirements.txt
    #   jsonpickle
    #   kombu
    #   pluggy
    #   pytest
    #   sqlalchemy
influxdb==5.2.1
    # via -r requirements.in
iniconfig==1.1.1
    # via pytest
jaeger-client==4.3.0
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
jsonext==0.4.2
    # via metadata-sdk
jsonpickle==2.0.0
    # via deepdiff
kafka-python==1.4.2
    # via -r requirements.in
kazoo==2.5.0
    # via
    #   -c ../upizza/requirements.txt
    #   metadata-sdk
kombu==4.6.11
    # via
    #   -c ../upizza/requirements.txt
    #   celery
    #   metadata-sdk
metadata_sdk==0.3.3
    # via -r requirements.in
numpy==1.19.5
    # via
    #   pandas
    #   pyarrow
opentracing==2.3.0
    # via
    #   -c ../upizza/requirements.txt
    #   django-opentracing
    #   jaeger-client
    #   opentracing-instrumentation
opentracing_instrumentation==3.3.1
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
packaging==20.9
    # via pytest
pandas==0.24.2
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
portalocker==2.3.0
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
    #   concurrent-log-handler
py==1.10.0
    # via pytest
pyarrow==0.16.0
    # via -r requirements.in
pycrypto==2.6.1
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
    #   metadata-sdk
pyhdfs==0.2.2
    # via -r requirements.in
pymysql==1.0.2
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
pyparsing==2.4.7
    # via packaging
pytest==6.2.4
    # via pytest-ordering
pytest-ordering==0.6
    # via -r requirements.in
python-crontab==2.5.1
    # via
    #   -c ../upizza/requirements.txt
    #   django-celery-beat
python-dateutil==2.8.1
    # via
    #   -c ../upizza/requirements.txt
    #   arrow
    #   influxdb
    #   pandas
    #   python-crontab
pytz==2021.1
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
    #   celery
    #   django
    #   django-timezone-field
    #   influxdb
    #   pandas
redis==2.10.5
    # via
    #   -r requirements.in
    #   django-redis
requests==2.22.0
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
    #   influxdb
    #   metadata-sdk
    #   pyhdfs
simplejson==3.17.2
    # via pyhdfs
six==1.15.0
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
    #   django-opentracing
    #   influxdb
    #   jsonext
    #   kazoo
    #   opentracing-instrumentation
    #   pyarrow
    #   python-dateutil
    #   thrift
    #   transitions
sqlalchemy==1.4.19
    # via metadata-sdk
sqlparse==0.4.1
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
    #   django
supervisor==4.2.1
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
tblib==1.7.0
    # via metadata-sdk
threadloop==1.0.2
    # via
    #   -c ../upizza/requirements.txt
    #   jaeger-client
thrift==0.13.0
    # via
    #   -c ../upizza/requirements.txt
    #   jaeger-client
toml==0.10.2
    # via pytest
tornado==5.1.1
    # via
    #   -c ../upizza/requirements.txt
    #   jaeger-client
    #   opentracing-instrumentation
    #   threadloop
transitions==0.8.5
    # via -r requirements.in
typing-extensions==3.7.4.3
    # via
    #   -c ../upizza/requirements.txt
    #   arrow
    #   importlib-metadata
urllib3==1.25.11
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
    #   elasticsearch
    #   requests
uwsgi==2.0.19
    # via
    #   -c ../upizza/requirements.txt
    #   -r requirements.in
vine==1.3.0
    # via
    #   -c ../upizza/requirements.txt
    #   amqp
    #   celery
wrapt==1.12.1
    # via
    #   -c ../upizza/requirements.txt
    #   opentracing-instrumentation
zipp==3.4.1
    # via
    #   -c ../upizza/requirements.txt
    #   importlib-metadata
zope.event==4.5.0
    # via
    #   -c ../upizza/requirements.txt
    #   gevent
zope.interface==5.4.0
    # via
    #   -c ../upizza/requirements.txt
    #   gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
